+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- upgrade: update spring-boot from v3.2.1 to v3.3.1
+
 ### Added
 - ability to send plain strings with `NostrTemplate`
