+++ docker/regtest/pg/docker-compose.yml
@@ -4,7 +4,7 @@ services:
   postgres:
     restart: always
     container_name: regtest_db
-    image: postgres:16.0-alpine3.18
+    image: postgres:16.2-alpine3.18
     volumes:
       - postgres-data:/postgresql_data
       - ./data/pg/init:/docker-entrypoint-initdb.d/
