+++ examples/nostr-relay-example-application/src/main/java/org/tbk/nostr/example/relay/domain/event/SqliteEventEntityService.java
@@ -4,7 +4,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.transaction.annotation.Transactional;
 import org.tbk.nostr.example.relay.NostrRelayExampleApplicationProperties;
-import org.tbk.nostr.example.relay.db.SupportedDatabaseType;
 import org.tbk.nostr.example.relay.impl.nip50.Nip50SearchTerm;
 import org.tbk.nostr.proto.Filter;
