+++ gradle.properties
@@ -3,4 +3,4 @@ org.gradle.configureondemand=true
 org.gradle.daemon=true
 org.gradle.dependency.verification.console=verbose
 org.gradle.jvmargs=-Xmx4g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
-org.gradle.parallel=true
+org.gradle.parallel=false
