+++ versions.gradle
@@ -16,7 +16,7 @@ ext {
     springdocOpenApiVersion = '2.1.0'
     checkstyleVersion = '10.3.4'
     findsecbugsPluginVersion = '1.12.0'
-    springBootGradlePluginVersion = '3.2.5'
+    springBootGradlePluginVersion = '3.3.1'
     springBootVersion = springBootGradlePluginVersion
     springShellVersion = '3.2.4'
     protobufGradleVersion = '0.9.4'
